requirements.txt
appnope==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
click==6.7
decorator==4.1.1
Flask==0.12.2
h5py==2.7.0
html5lib==0.9999999
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
Keras==2.0.6
Markdown==2.6.8
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.2.1
Theano==0.9.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
